Add files via upload
TF-IDF (term frequency, inverse document frequency) is a  statistical text representation technique  in natural language processing and information retrieval. .It  is one of the most popular text vectorizers. 

    This IDF metric can be calculated by taking the total number of documents, dividing it by the number of documents that contain a word, and calculating the logarithm.
    So, if the word is very common and appears in many documents, this number will approach 0. Otherwise, it will approach 1.
    TF being a raw count of instances a word appears in a document.
Multiplying these two numbers results in the TF-IDF score of a word in a document. The higher the score, the more relevant that word is in that particular document.